Base class for cached data based MOs
This abstract class provides some of the necessary plumbing for managed
objects (MOs) that operate mostly on data that can be cached in the MO.

It is not meant as an appropriate base for MOs that operate on data that is
volatile from the device management application's point of view (e.g. sensor
measurements, log files) and/or heavily need to trigger immediate local action
based on commands (get/set) from the backend.