MOTree initial implementation
Initial / minimal implementation of the MOTree class representing
the structure of all MO instances held inside the client libary.

The MO_StaticData class contains one of the simplest MO implementations possible
It represents a passive data MO. Node values can be read and updated from both
the local application and the DM server,  but no side effect or additional action
will be triggered by a DM command, except the change in value of the node content.